ai-planner-monorepo/services/api-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.28.1
